Adds crash post client
