+++ spec/config.ts
@@ -1,5 +1,6 @@
 export const config = {
     host: 'https://app.bugsplat.com',
     email: 'fred@bugsplat.com',
-    password: <string>process.env.FRED_PASSWORD
+    password: <string>process.env.FRED_PASSWORD,
+    database: 'fred'
 };
