+++ src/common/client/index.ts
@@ -1,2 +1,3 @@
 export { ApiClient, BugSplatResponse } from './api-client';
-export { BugSplatApiClient } from './bugsplat-api-client';
+export { BugSplatApiClient } from './bugsplat-api-client/bugsplat-api-client';
+export { S3ApiClient } from './s3-api-client/s3-api-client';
