+++ src/common/file/bugsplat-file.ts
@@ -1,0 +1,9 @@
+import type fs from 'fs';
+
+export class BugSplatFile {
+    constructor(
+        public readonly name: string,
+        public readonly size: number,
+        public readonly file: File | fs.ReadStream
+    ) { }
+}
