+++ src/common/index.ts
@@ -1,3 +1,4 @@
 export * from './client';
 export * from './data';
+export { BugSplatFile } from './file/bugsplat-file';
 export { Environment } from './environment';
