+++ src/post/crash-type.ts
@@ -1,0 +1,4 @@
+export enum CrashType {
+    native = 'Windows.Native',
+    dotnet = 'Windows.NET'
+}
