+++ src/symbols/symbols-api-client/symbols-api-client.ts
@@ -1,15 +1,16 @@
-import type fs from 'fs';
+import { ApiClient, BugSplatFile, BugSplatResponse, S3ApiClient } from '@common';
 import { lastValueFrom, timer } from 'rxjs';
-import { ApiClient, BugSplatResponse } from '../../common';
-import { S3ApiClient } from '../s3-api-client/s3-api-client';
 
 export class SymbolsApiClient {
 
     private readonly route = '/api/symbols';
 
+    private _s3ApiClient: S3ApiClient;
     private _timer = timer;
 
-    constructor(private _client: ApiClient) { }
+    constructor(private _client: ApiClient) {
+        this._s3ApiClient = new S3ApiClient();
+    }
 
     async delete(
         database: string,
@@ -41,7 +42,7 @@ export class SymbolsApiClient {
         database: string,
         application: string,
         version: string,
-        files: Array<{ name: string, size: number, file: File | fs.ReadStream }>
+        files: Array<BugSplatFile> 
     ): Promise<Array<BugSplatResponse>> {
         const promises = files
             .map(async (file) => {
@@ -55,8 +56,7 @@ export class SymbolsApiClient {
                     name
                 );
     
-                const s3Client = new S3ApiClient();
-                const response = s3Client.uploadFileToPresignedUrl(presignedUrl, file.file, size);
+                const response = await this._s3ApiClient.uploadFileToPresignedUrl(presignedUrl, file);
                 await lastValueFrom(this._timer(1000));
 
                 return response;
